% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_kraken_order}
\alias{add_kraken_order}
\title{add order}
\usage{
add_kraken_order(account, assetpair, direction, type, price, volume)
}
\arguments{
\item{account}{a valid account (see kraken_query())}

\item{assetpair}{a valid assetpair}

\item{direction}{trade direction (buy or sell)}

\item{type}{an ordertype}

\item{price}{a price for the order}

\item{volume}{volume to trade}
}
\value{
result of AddOrder query on Kraken
}
\description{
add order
}
