% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cmc.R
\name{api_cmc_top_n}
\alias{api_cmc_top_n}
\title{Extract top_n currencies from CoinMarketCap}
\usage{
api_cmc_top_n(top_n = 40)
}
\arguments{
\item{top_n}{number of currencies to list (<100)}
}
\value{
top n currencies by market-cap
}
\description{
Extract top_n currencies from CoinMarketCap
}
\examples{
\dontrun{
api_cmc_top_n()
}
}
