% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_binance.R
\name{binance_klines}
\alias{binance_klines}
\title{Get kline/candlestick data from Binance}
\usage{
binance_klines(symbol, interval, limit = 500, start_time, end_time)
}
\arguments{
\item{symbol}{string}

\item{interval}{enum}

\item{limit}{int}

\item{start_time}{POSIX timestamp}

\item{end_time}{POSIX timestamp}
}
\value{
data.table
}
\description{
Get kline/candlestick data from Binance
}
\examples{
\dontrun{
binance_klines('ETHUSDT')
binance_klines('ETHUSDT', interval = '1h', limit = 24*7)
binance_klines('ETHUSDT', interval = '1h',
    start_time = as.POSIXct('2018-01-01'), end_time = as.POSIXct('2018-01-08'))
}
}
