% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{calculate_retracement}
\alias{calculate_retracement}
\title{Calculate Retracement Ratio}
\usage{
calculate_retracement(data)
}
\arguments{
\item{data}{A dataframe with at least \code{date} and \code{price} columns.}
}
\value{
A dataframe with an additional \code{retrace_ratio} column.
}
\description{
Computes the retracement ratio for a given cryptocurrency.
}
