% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_coins_df.R
\name{cg_coins_df}
\alias{cg_coins_df}
\title{Build a custom data frame of coins}
\usage{
cg_coins_df(per_page, pages)
}
\arguments{
\item{per_page}{\code{integer} Symbols per page; from 1 - 250.}

\item{pages}{\code{integer} Number of pages to return.}
}
\value{
Returns a "data.frame".
}
\description{
A wrapper for \code{cg_get_coins_markets(...)} which is limited to 250
symbols per page. This function allows a custom data frame to be build by
progressively binding data frames given the function's parameters.
}
\details{
The \code{cg_get_coins_markets(...)} function is by default ordered according
to market cap and in descending order.
}
\section{Reference}{
 See
\code{\link{cg_get_coins_markets}}
}

\examples{
cg_coins_df(per_page = 250, pages = 4)

}
