% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_get_global.R
\name{cg_get_global}
\alias{cg_get_global}
\title{List cryptocurrency data}
\usage{
cg_get_global()
}
\value{
Returns a "list".
}
\description{
Get global cryptocurrency data
}
\section{Reference}{
 See
\code{/global}
in the
\href{https://www.coingecko.com/api/documentations/v3}{API documentation}.
}

\examples{
cg_get_global()

}
