% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_get_ping.R
\name{cg_get_ping}
\alias{cg_get_ping}
\title{Check API server status}
\usage{
cg_get_ping()
}
\value{
A "list"
}
\description{
Get CoinGecko API server status.
}
\section{Reference}{
 See
\code{/ping}
in the
\href{https://www.coingecko.com/api/documentations/v3}{API documentation}.
}

\examples{
cg_get_ping()

}
