% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{check_credentials_binance}
\alias{check_credentials_binance}
\title{Check if Binance credentials were set previously}
\usage{
check_credentials_binance()
}
\value{
fail on missing credentials
}
\description{
Check if Binance credentials were set previously
}
\keyword{internal}
