% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{check_credentials_kraken}
\alias{check_credentials_kraken}
\title{Check if Kraken credentials were set previously}
\usage{
check_credentials_kraken()
}
\value{
fail on missing credentials
}
\description{
Check if Kraken credentials were set previously
}
\keyword{internal}
