% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{check_credentials_nomics}
\alias{check_credentials_nomics}
\title{Check if Nomics credentials were set previously}
\usage{
check_credentials_nomics()
}
\value{
fail on missing credentials
}
\description{
Check if Nomics credentials were set previously
}
\keyword{internal}
