% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{close_kraken_open_orders}
\alias{close_kraken_open_orders}
\title{Close all open orders or a selection of order ids}
\usage{
close_kraken_open_orders(ids = NULL)
}
\arguments{
\item{refids}{vector of refids to close (optional)}
}
\value{
nothing
}
\description{
Close all open orders or a selection of order ids
}
