% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{close_open_orders}
\alias{close_open_orders}
\title{Close all open orders}
\usage{
close_open_orders(refids = NULL)
}
\arguments{
\item{refids}{vector of refids to close (optional)}
}
\value{
nothing
}
\description{
Close all open orders
}
