% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials_binance}
\alias{credentials_binance}
\title{Sets the API key and secret to interact with the Binance API}
\usage{
credentials_binance(key, secret)
}
\arguments{
\item{key}{string}

\item{secret}{string}
}
\description{
Sets the API key and secret to interact with the Binance API
}
\examples{
\dontrun{
credentials_binance('foo', 'bar')
}
}
