% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials_kraken}
\alias{credentials_kraken}
\title{Sets the API key and secret to interact with the Kraken API}
\usage{
credentials_kraken(key, secret)
}
\arguments{
\item{key}{string}

\item{secret}{string}
}
\description{
Sets the API key and secret to interact with the Kraken API
}
\examples{
\dontrun{
credentials_kraken('foo', 'bar')
}
}
