% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials_nomics}
\alias{credentials_nomics}
\title{Sets the API key and secret to interact with the Nomics API}
\usage{
credentials_nomics(key)
}
\arguments{
\item{key}{string}
}
\description{
Sets the API key and secret to interact with the Nomics API
}
\examples{
\dontrun{
credentials_nomics('foo')
}
}
