% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_long_to_wide}
\alias{data_long_to_wide}
\title{Convert long dataframe to a wide format}
\usage{
data_long_to_wide(long)
}
\arguments{
\item{long}{a dataframe in long format}
}
\value{
a dataframe
}
\description{
Convert long dataframe to a wide format
}
