% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_wide_to_long}
\alias{data_wide_to_long}
\title{Convert a wide dataframe to a long format}
\usage{
data_wide_to_long(wide)
}
\arguments{
\item{wide}{a dataframe in wide format}
}
\value{
a dataframe
}
\description{
Convert a wide dataframe to a long format
}
