% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_rfc3339}
\alias{from_rfc3339}
\title{Convert an RFC3339 date to an R date}
\usage{
from_rfc3339(date)
}
\arguments{
\item{date}{an RFC3339 date}
}
\value{
a POSIXct representation of date
}
\description{
Convert an RFC3339 date to an R date
}
