% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_api_limit}
\alias{get_api_limit}
\title{Get number of API calls used in last hour or last second.}
\usage{
get_api_limit(time = "hour")
}
\arguments{
\item{time}{character. Either "hour" or "second". Defaults to "hour".}
}
\value{
A list of two tibbles detailing calls made and calls remaining against
the API limit
}
\description{
\code{get_api_limit} returns a list of two tibbles detailing API calls made
and API calls remaining against second and hour rate limits.
}
\examples{
\dontrun{
# Get hour limit
get_api_limit()

# Get second limit
get_api_limit(time = "second")
}
}
\references{
https://www.cryptocompare.com/api#requests
}
