% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcap_price.R
\encoding{UTF-8}
\name{get_cmc_id}
\alias{get_cmc_id}
\title{get_cmc_id}
\usage{
get_cmc_id(txt)
}
\arguments{
\item{txt}{string text to lookup in mapping object}
}
\value{
numberic id
}
\description{
Function to get CoinMarketCap id based on a crypto symbol
}
\examples{
get_cmc_id("luna")

}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
}
\concept{Lookup corresponding id to use to download prices for the correct crypto}
