% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcap_price.R
\encoding{UTF-8}
\name{get_cmc_map}
\alias{get_cmc_map}
\title{get_cmc_map}
\usage{
get_cmc_map()
}
\value{
data.table with mapping
}
\description{
Mapping from CoinMarketCap id to crypto symbol
}
\examples{
get_cmc_map()

}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
}
\concept{Mapping from CoinMarketCap id to crypto symbol}
