% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmarketcap_price.R
\encoding{UTF-8}
\name{get_cmc_price}
\alias{get_cmc_price}
\title{get_cmc_price}
\usage{
get_cmc_price(id = NULL, time_start = "2018-01-01", time_end = Sys.Date())
}
\arguments{
\item{id}{numeric id or string symbol}

\item{time_start}{starting date}

\item{time_end}{end date, defaulting to today}
}
\value{
data.table with symbol, OHLC, volume, and name for the input id
}
\description{
Get prices from the CoinMarketCap web api
}
\examples{
get_cmc_price(4172)
get_cmc_price("dot")

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get prices from the CoinMarketCap web api}
