% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_asset_platforms}
\alias{get_coingecko_asset_platforms}
\title{get_coingecko_asset_platforms}
\usage{
get_coingecko_asset_platforms()
}
\arguments{
\item{None}{}
}
\value{
vector of asset_platforms
}
\description{
Get list of asset_platforms from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_asset_platforms()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get asset_platforms from the CoinGecko web api}
