% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_coin_categories}
\alias{get_coingecko_coin_categories}
\title{get_coingecko_coin_categories}
\usage{
get_coingecko_coin_categories()
}
\arguments{
\item{None}{}
}
\value{
data.frame of coin categories
}
\description{
Get list of coin categories from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_coin_categories()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get coin categories from the CoinGecko web api}
