% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_currencies}
\alias{get_coingecko_currencies}
\title{get_coingecko_currencies}
\usage{
get_coingecko_currencies()
}
\arguments{
\item{None}{}
}
\value{
vector of supported currencies
}
\description{
Get list of supported vs currencies from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_currencies()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get status from the CoinGecko web api}
