% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_global}
\alias{get_coingecko_global}
\title{get_coingecko_global}
\usage{
get_coingecko_global()
}
\arguments{
\item{None}{}
}
\value{
data.frame of cryptocurrency global data
}
\description{
Get cryptocurrency global data from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_global()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get cryptocurrency global data from the CoinGecko web api}
