% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_history}
\alias{get_coingecko_history}
\title{get_coingecko_history}
\usage{
get_coingecko_history(id = NULL, cur = "usd", days = "max")
}
\arguments{
\item{id}{string Id}

\item{cur}{string currency (defaults to 'usd')}

\item{days}{integer (1/7/14/30/90/180/365) or string "max"}
}
\value{
data.table of date, price, market_cap, volume, Id, and epoch (unix time)
}
\description{
Get historical data from the CoinGecko web api for a given Id (not Symbol)
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_history("cardano", days = 30)
tmp <- get_coingecko_history("bitcoin", days = "max")

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get historical prices from the CoinGecko web api}
