% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_id_from_symbol}
\alias{get_coingecko_id_from_symbol}
\title{get_coingecko_id_from_symbol}
\usage{
get_coingecko_id_from_symbol(tag)
}
\arguments{
\item{txt}{string text to lookup in mapping object}
}
\value{
numberic id
}
\description{
Function to get CoinGecko id based on a crypto symbol
}
\examples{
get_coingecko_id_from_symbol("luna")
get_coingecko_id_from_symbol("LUNA")

}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
}
\concept{Lookup corresponding id to use to download prices for the correct crypto}
