% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_map}
\alias{get_coingecko_map}
\title{get_coingecko_map}
\usage{
get_coingecko_map()
}
\value{
data.frame with mapping
}
\description{
Mapping from CoinGecko id to crypto symbol
}
\examples{
get_coingecko_map()

}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
}
\concept{Mapping from CoinGecko id to crypto symbol}
