% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_name}
\alias{get_coingecko_name}
\title{get_coingecko_name}
\usage{
get_coingecko_name(txt)
}
\arguments{
\item{txt}{string to search for in the CoinGecko mapping table}
}
\value{
data.frame with all observations that match the input string
}
\description{
Function to return the crypto's matching the search criteria
}
\examples{
get_coingecko_name("luna")

}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
}
\concept{Function to return the crypto's matching the search criteria}
