% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_ohlc}
\alias{get_coingecko_ohlc}
\title{get_coingecko_ohlc}
\usage{
get_coingecko_ohlc(id = NULL, cur = "usd", days = "max")
}
\arguments{
\item{id}{string Id}

\item{cur}{string currency (defaults to 'usd')}

\item{days}{integer (1/7/14/30/90/180/365) or string "max"}
}
\value{
data.frame of date, open, high, low, close, Id, and epoch (unix time)
}
\description{
Get OHLC candle data from the CoinGecko web api for a given Id (not Symbol)
Candle's body:
1 - 2 days: 30 minutes
3 - 30 days: 4 hours
31 and before: 4 days
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_ohlc("cardano", days = 30)
tmp <- get_coingecko_ohlc("bitcoin", days = "max")

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get prices from the CoinGecko web api}
