% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_ping}
\alias{get_coingecko_ping}
\alias{.isnull}
\title{get_coingecko_ping}
\usage{
.isnull(x, y)

get_coingecko_ping()
}
\arguments{
\item{None}{}
}
\value{
string "(V3) To the Moon!" means everything is okay on the CoinGecko side.
}
\description{
Get status from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_ping()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}

Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get status from the CoinGecko web api}
