% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_price}
\alias{get_coingecko_price}
\title{get_coingecko_price}
\usage{
get_coingecko_price(id = NULL, cur = "usd")
}
\arguments{
\item{id}{string Id}

\item{cur}{string currency (defaults to 'usd')}
}
\value{
float price
}
\description{
Get LATEST price from the CoinMarketCap web api for a given Id (not Symbol)
https://www.coingecko.com/en/api/documentation
}
\examples{
get_coingecko_price("cardano")
get_coingecko_price("bitcoin")

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get prices from the CoinGecko web api}
