% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_public_treasury}
\alias{get_coingecko_public_treasury}
\title{get_coingecko_public_treasury}
\usage{
get_coingecko_public_treasury()
}
\arguments{
\item{None}{}
}
\value{
data.frame of treasury holdings by public companies
}
\description{
Get public treasury holdings from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_public_treasury()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get public treasury holdings from the CoinGecko web api}
