% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coingecko_price.R
\encoding{UTF-8}
\name{get_coingecko_trending}
\alias{get_coingecko_trending}
\title{get_coingecko_trending}
\usage{
get_coingecko_trending()
}
\arguments{
\item{None}{}
}
\value{
data.frame of coin categories
}
\description{
Get Top-7 trending coins on CoinGecko as searched by users in the last 24 hours (Ordered by most popular first) from the CoinGecko web api
https://www.coingecko.com/en/api/documentation
}
\examples{
tmp <- get_coingecko_trending()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get Top-7 trending coins on CoinGecko as searched by users in the last 24 hours (Ordered by most popular first) from the CoinGecko web api}
