% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_coins}
\alias{get_coins}
\title{List all available coins}
\usage{
get_coins()
}
\value{
A tibble containing details for each coin available through the API
}
\description{
\code{get_coins} provides general information about all coins available through
the API.
}
\examples{
\dontrun{
coins <- get_coins()
head(coins)
}
}
\references{
https://www.cryptocompare.com/api#-api-data-coinlist-
}
