% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptofees.R
\encoding{UTF-8}
\name{get_cyrptofees_fees}
\alias{get_cyrptofees_fees}
\title{get_cyrptofees_fees}
\usage{
get_cyrptofees_fees()
}
\arguments{
\item{None}{}
}
\value{
list of protocols, fees, and events
}
\description{
Get crypto fees paid from the Cryptofees.info web api
https://cryptofees.info/api-docs
}
\examples{
tmp <- get_cyrptofees_fees()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get crypto fees paid from the Cryptofees.info web api}
