% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_day_avg}
\alias{get_day_avg}
\title{Get averge daily price for a given currency pair.}
\usage{
get_day_avg(fsym, tsym, end_time = Sys.time(), exchange = "CCCAGG",
  avg_type = "HourVWAP", utc_hour_diff = 0, sign = FALSE,
  try_conversion = TRUE, app_name = NULL)
}
\arguments{
\item{fsym}{character. 3 letter name for coin to retreive price history for.}

\item{tsym}{character. 3 letter name for how price should be reported for
\code{fsym}.}

\item{end_time}{character or POSIXct or POSIXt. Indicates the date data should
be retrieved for. Defaults to \code{Sys.time()}.}

\item{exchange}{character. Exchange to query. Defaults to 'CCCAGG'.}

\item{avg_type}{character. Indicates the type of calculation to use. One of
"HourVWAP", "MidHighLow" or "VolFVolT".}

\item{utc_hour_diff}{numeric. Hour difference between current timezone and UTC.
The API returns results in UTC by default. Valid values are between -12 and 14.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{try_conversion}{logical. Should BTC conversion be used fsym is not trading
in tsym on specified exchange? Defaults to TRUE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing the date, fsym, tsym, and daily average price
}
\description{
\code{get_day_avg} returns the daily average for a given currency pair calculated
in one of 3 ways.
}
\examples{
\dontrun{
# Get today's daily average of BTC in USD
get_day_avg("BTC", "USD")
}

}
\references{
\url{https://min-api.cryptocompare.com}
}
