% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_exchange.R
\name{get_exchanges}
\alias{get_exchanges}
\title{Get Exchange Data}
\usage{
get_exchanges()
}
\value{
A dataframe containing exchange ID, name, country, trading volume, and trust score.
}
\description{
Retrieve information about cryptocurrency exchanges, including trading volume, trust score, and more.
}
