% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_api.R
\encoding{UTF-8}
\name{get_kraken}
\alias{get_kraken}
\title{get_kraken}
\usage{
get_kraken(
  method,
  pair = "",
  since = "2020-01-01",
  interval = 1,
  info = "info",
  txid = "",
  asset = "all",
  type = "all",
  ofs = 1,
  enddate = Sys.Date(),
  timeout = 60,
  docalcs = "true",
  consolidation = "market",
  trades = "",
  ...
)
}
\arguments{
\item{method}{string one of ("Ticker","Trades","Depth","OHLC","Spread","AssetPairs","Assets","Time","Balance","Ledgers","OpenOrders","AddOrder","CalcelOrder","CalcelAll")}

\item{pair}{string currency pair}

\item{since}{numeric UNIX time for start of query}

\item{interval}{numeric, default of 1}

\item{count}{numeric how many observations to return, used for `Depth`}
}
\value{
data.table or json
}
\description{
Helper function called by get_kraken
}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
#' @examples
get_kraken("Ticker", pair = "DOTUSD")
get_kraken("Trades", pair = "DOTUSD")
get_kraken("Depth", pair = "DOTUSD")
get_kraken("OHLC", pair = "DOTUSD")
get_kraken("Spread", pair = "DOTUSD")
get_kraken("AssetPairs", pair = "DOTUSD")
get_kraken("Assets")
get_kraken("Time")
get_kraken("Balance")
get_kraken("Ledgers")
get_kraken("OpenOrders")
}
\concept{Helper function called by get_kraken}
