% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken_api.R
\encoding{UTF-8}
\name{get_kraken_url}
\alias{get_kraken_url}
\title{get_kraken_url}
\usage{
get_kraken_url(url, params = NULL, key = NULL, secret = NULL)
}
\value{
data.table with mapping

Please see https://support.kraken.com/hc/en-us/articles/360029054811-What-is-the-authentication-algorithm-for-private-endpoints-
}
\description{
Helper function called by get_kraken
}
\author{
Roger J. Bos, CFA, \email{me@rogerjbos.com}
}
\concept{Helper function called by get_kraken}
