% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_news_providers}
\alias{get_news_providers}
\title{Get details on news providers integrated on cryptocompare.}
\usage{
get_news_providers(sign = FALSE, app_name = NULL)
}
\arguments{
\item{sign}{logical. Should the server sign the response? Defaults to FALSE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containining simple details about integrated news providers.
}
\description{
\code{get_news_providers} returns details about news providers that are
integrated with the cryptocompare website.
}
\references{
\url{https://min-api.cryptocompare.com}
}
