% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_pair_snapshot}
\alias{get_pair_snapshot}
\title{Get data for a currency pair}
\usage{
get_pair_snapshot(fsym, tsym)
}
\arguments{
\item{fsym}{character. 3 letter name for coin to retreive.}

\item{tsym}{character. 3 letter name for how details should be reported for
\code{fsym}.}
}
\value{
A list of 3 containing the following:
\describe{
  \item{coin_data}{A tibble containing general coin data}
  \item{aggregated_data}{A tibble containing aggregate data about the pair}
  \item{exchange_data}{A tibble containing exchange data for each available
       exchange and the provided pair}
}
}
\description{
\code{get_pair_snapshot} provides aggregated and individual data for the
provided currency pair in every exchange available.
}
\examples{
\dontrun{
# Get data on BTC and USD
get_pair_snapshot("BTC", "USD")
}

}
\references{
\url{https://www.cryptocompare.com/api#-api-data-coinsnapshot-}
}
