% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_price}
\alias{get_price}
\title{Get price data for one or more coin(s)}
\usage{
get_price(fsyms, tsyms, exchange = "CCCAGG", sign = FALSE,
  try_conversion = TRUE, app_name = NULL)
}
\arguments{
\item{fsyms}{character. 3 letter name(s) for coins to retreive price details for.}

\item{tsyms}{character. 3 letter name(s) for how price should be reported for
\code{fsyms}.}

\item{exchange}{character. Exchange to query. Defaults to 'CCCAGG'.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{try_conversion}{logical. Should BTC conversion be used fsym is not trading
in tsym on specified exchange? Defaults to TRUE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing price details about selected coin(s).
}
\description{
\code{get_price} provides price information about selected coins in terms
of other coins.
}
\examples{
\dontrun{
# Get BTC price in USD
get_price("BTC", "USD")

# Get BTC and ETH price in USD and EUR
get_price(c("BTC", "ETH"), c("USD", "EUR"))
}

}
\references{
https://www.cryptocompare.com/api#-api-data-price-
}
