% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_price_avg}
\alias{get_price_avg}
\title{Get weighted average price of coin pair based on any number of exchanges.}
\usage{
get_price_avg(fsym, tsym, exchanges, sign = FALSE, try_conversion = TRUE,
  app_name = NULL)
}
\arguments{
\item{fsym}{character. 3 letter name for coin to retreive price history for.}

\item{tsym}{character. 3 letter name for how price should be reported for
\code{fsym}.}

\item{exchanges}{character. Exchanges to build weighted average from.
Defaults to 'CCCAGG'.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{try_conversion}{logical. Should BTC conversion be used fsym is not trading
in tsym on specified exchange? Defaults to TRUE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing price details as a volume weighted average of the
provided exchanges.
}
\description{
\code{get_price_avg} returns the current trading details
(price, vol, open, high, low, etc) of a given coin pair as a volume weighted
average based on provided exchanges.
}
\examples{
\dontrun{
get_price_avg("BTC", "USD", c("Coinbase", "Kraken", "Bitfinex"))
}

}
\references{
\url{https://min-api.cryptocompare.com}
}
