% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_price_details}
\alias{get_price_details}
\title{Get full price details for one or more coin(s)}
\usage{
get_price_details(fsyms, tsyms, exchange = "CCCAGG", sign = FALSE,
  app_name = NULL, try_conversion = TRUE)
}
\arguments{
\item{fsyms}{character. 3 letter name(s) for coins to retreive price details for.}

\item{tsyms}{character. 3 letter name(s) for how price should be reported for
\code{fsyms}.}

\item{exchange}{character. Exchange to query. Defaults to 'CCCAGG'.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}

\item{try_conversion}{logical. Should conversion be used if exact comparison
isn't available? Defaults to TRUE.}
}
\value{
A tibble containing comprehensive price details for \code{fsyms} in
terms of \code{tsyms}.
}
\description{
\code{get_price_details} is Like code{\link{get_price}} but provides a more
thorough summary of price data by including details like 24 volume and highs
and lows.
}
\references{
https://www.cryptocompare.com/api#-api-data-price-
}
