% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_social}
\alias{get_social}
\title{Gets social data for a given coin id.}
\usage{
get_social(id)
}
\arguments{
\item{id}{integer. Coin id. Can be retreived using \code{\link{get_coins}}.}
}
\description{
\code{get_social} returns a list containing several tibbles that provide
social details for the specified coin.
}
\examples{
\dontrun{
# Get bitcoin social data
get_social(1182)
}

}
\references{
\url{https://www.cryptocompare.com/api#-api-data-socialstats-}
}
