% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_top_coins_by_volume}
\alias{get_top_coins_by_volume}
\title{Get top coins by total volume across all markets.}
\usage{
get_top_coins_by_volume(tsym, limit = 10, page = 0, sign = FALSE,
  app_name = NULL)
}
\arguments{
\item{tsym}{character. 3 letter name for coin to get top volumes for.}

\item{limit}{numeric. Number of records to return. Defaults to 10.}

\item{page}{numeric. Page of results to return. Defaults to 0.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing details about the top coins by volume for the
specified coin.
}
\description{
\code{get_top_coins_by_volume} returns details about the top coins by volume
for the specified coin (\code{tsym}).
}
\examples{
\dontrun{
get_top_coins_by_volume("USD")
}

}
\references{
\url{https://min-api.cryptocompare.com}
}
