% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_top_exchanges_full}
\alias{get_top_exchanges_full}
\title{Get top exchanges by volume along with additional details.}
\usage{
get_top_exchanges_full(fsym, tsym, limit = 5, sign = FALSE,
  app_name = NULL)
}
\arguments{
\item{fsym}{character. 3 letter name for coin to retreive price history for.}

\item{tsym}{character. 3 letter name for how price should be reported for
\code{fsym}.}

\item{limit}{numeric. Maximum number of returned exchanges. Between 1 and 50.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A named list containing exchanges, coin_info, and aggregated_data.
}
\description{
\code{get_top_exchanges_full} retrieves top exchanges for the given coin pair
by volume along with additional coin and exchange details.
}
\examples{
\dontrun{
get_top_exchanges_full("BTC", "USD")
}
}
\references{
\url{https://min-api.cryptocompare.com}
}
