% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_top_pairs}
\alias{get_top_pairs}
\title{Get top currencies by volume for a given currency.}
\usage{
get_top_pairs(fsym, limit = 5, sign = FALSE)
}
\arguments{
\item{fsym}{character. Three letter symbol for currency of interest.}

\item{limit}{integer. Number of top pairs to return. Min of 1, max of 2000.
Defaults to 5.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE.}
}
\value{
A tibble containing the top currencies by volume for the specified
currency
}
\description{
\code{get_top_pairs} returns a tibble containing the top currency pairs by
volume.
}
\examples{
\dontrun{
# Get top 10 currencies for bitcoin
get_top_pairs("BTC", limit = 10)

# Get top 20 currencies for ethereum
get_top_pairs("ETH", limit = 20)
}

}
\references{
\url{https://www.cryptocompare.com/api#-api-data-toppairs-}
}
