% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_top_volumes}
\alias{get_top_volumes}
\title{Get top coins by volume for the provided currency.}
\usage{
get_top_volumes(tsym, limit = 20, sign = FALSE, app_name = NULL)
}
\arguments{
\item{tsym}{character. 3 letter name for coin to get top volumes for.}

\item{limit}{numeric. Maximum number of records to return. Defaults to 20.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing details about the top coins by volume for the
specified coin.
}
\description{
\code{get_top_volumes} returns details about the top coins by volume for the
specified currency.
}
\examples{
\dontrun{
# Get top volumes traded into USD
get_top_volumes("USD")
}

}
\references{
\url{https://min-api.cryptocompare.com}
}
