% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_series}
\alias{index_series}
\title{Index a timeseries at a given point}
\usage{
index_series(wide, lookback)
}
\arguments{
\item{wide}{a dataframe in wide format}

\item{lookback}{period to look back}
}
\value{
an indexed time-series
}
\description{
Index a timeseries at a given point
}
