% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{key_binance}
\alias{key_binance}
\title{Look up Binance API key stored in the environment}
\usage{
key_binance()
}
\value{
string
}
\description{
Look up Binance API key stored in the environment
}
\keyword{internal}
