% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{key_kraken}
\alias{key_kraken}
\title{Look up Kraken API key stored in the environment}
\usage{
key_kraken()
}
\value{
string
}
\description{
Look up Kraken API key stored in the environment
}
\keyword{internal}
