% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{key_nomics}
\alias{key_nomics}
\title{Look up Nomics API key stored in the environment}
\usage{
key_nomics()
}
\value{
string
}
\description{
Look up Nomics API key stored in the environment
}
\keyword{internal}
