% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_add_order}
\alias{kraken_add_order}
\title{Add standard order}
\usage{
kraken_add_order(pair, type, ordertype, price, price2, volume,
  validate = TRUE)
}
\arguments{
\item{pair}{asset pair}

\item{type}{type of order (buy/sell)}

\item{ordertype}{order type:}

\item{price}{price (optional.  dependent upon ordertype)}

\item{price2}{secondary price (optional.  dependent upon ordertype)}

\item{volume}{order volume in lots}

\item{validate}{validate inputs only.  do not submit order (optional)}
}
\value{
void
}
\description{
Add standard order
}
