% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_add_order_internal}
\alias{kraken_add_order_internal}
\title{Add standard order}
\usage{
kraken_add_order_internal(args)
}
\arguments{
\item{args}{list of arguments to include in API call}
}
\value{
order description info and array of transaction ids for order (if order was added successfully)
}
\description{
Add standard order
}
\keyword{internal}
