% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_add_order_until_complete}
\alias{kraken_add_order_until_complete}
\title{Add orders until rebalanced}
\usage{
kraken_add_order_until_complete(pair, type, volume)
}
\arguments{
\item{pair}{an assetpair}

\item{type}{a type of order (buy or sell)}

\item{volume}{volume to execute}
}
\value{
nothing
}
\description{
Add orders until rebalanced
}
