% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_allocation_current}
\alias{kraken_allocation_current}
\title{Get current and target allocation}
\usage{
kraken_allocation_current(target_allocation)
}
\arguments{
\item{target_allocation}{}
}
\description{
Get current and target allocation
}
