% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_cancel_open_order}
\alias{kraken_cancel_open_order}
\title{Cancel open order}
\usage{
kraken_cancel_open_order(txid)
}
\arguments{
\item{txid}{transaction id}
}
\value{
list of number of orders canceled and pending information
}
\description{
Cancel open order
}
