% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_cancel_open_orders}
\alias{kraken_cancel_open_orders}
\title{Cancel open orders or a selection of order ids}
\usage{
kraken_cancel_open_orders(ids)
}
\arguments{
\item{ids}{vector or list of txid's to close (optional)}
}
\value{
list of return values from kraken_cancel_open_order()
}
\description{
Cancel open orders or a selection of order ids
}
