% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kraken_candles}
\alias{kraken_candles}
\title{get candle data}
\usage{
kraken_candles(pairs, interval = 1440)
}
\arguments{
\item{pairs}{list of assetpair names or altnames to retrieve OHLC data.}

\item{interval}{integer corresponding to candle interval}
}
\value{
a dataframe with OHLC data
}
\description{
get candle data
}
