% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_get_closed_orders}
\alias{kraken_get_closed_orders}
\title{Get closed orders}
\usage{
kraken_get_closed_orders()
}
\arguments{
\item{userref}{restrict results to given user reference id (optional)}
}
\value{
dataframe of order info
}
\description{
Get closed orders
}
