% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_get_deposit_address}
\alias{kraken_get_deposit_address}
\title{Get deposit addresses}
\usage{
kraken_get_deposit_address(aclass = NULL, asset, method, new = NULL)
}
\arguments{
\item{aclass}{(optional) asset class}

\item{asset}{(required) asset being deposited}

\item{method}{(required) name of the deposit method}

\item{new}{whether or not to generate a new address (optional. default = false)}
}
\value{
associative array of deposit addresses
}
\description{
Get deposit addresses
}
