% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_get_deposit_methods}
\alias{kraken_get_deposit_methods}
\title{Get deposit methods}
\usage{
kraken_get_deposit_methods(aclass = NULL, asset)
}
\arguments{
\item{aclass}{(optional) asset class}

\item{asset}{(required) asset being deposited}
}
\value{
associative array of deposit methods
}
\description{
Get deposit methods
}
