% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_get_open_orders}
\alias{kraken_get_open_orders}
\title{Get open orders}
\usage{
kraken_get_open_orders(userref = NULL)
}
\arguments{
\item{userref}{(optional) a given 'userref' to query for (default = NULL)}
}
\value{
array of order info in open array with txid as the key
}
\description{
Get open orders
}
