% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_get_withdrawal_info}
\alias{kraken_get_withdrawal_info}
\title{Get withdrawal information}
\usage{
kraken_get_withdrawal_info(aclass = NULL, asset, key, amount)
}
\arguments{
\item{aclass}{(optional) asset class}

\item{asset}{asset being withdrawn}

\item{key}{withdrawal key name, as set up on your account}

\item{amount}{amount to withdraw}
}
\value{
associative array of withdrawal info:
}
\description{
Get withdrawal information
}
