% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_lastprice}
\alias{kraken_lastprice}
\title{Get a pair's last available price}
\usage{
kraken_lastprice(pair)
}
\arguments{
\item{pair}{a valid Kraken asset pair.}
}
\value{
last available price for a given pair
}
\description{
Get a pair's last available price
}
