% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_ledger}
\alias{kraken_ledger}
\title{get ledger information}
\usage{
kraken_ledger(simple = TRUE)
}
\arguments{
\item{simple}{(optional) Get 'simple' ledger or not (default = TRUE)}
}
\value{
returns the ledger dataframe
}
\description{
get ledger information
}
