% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_ohlc}
\alias{kraken_ohlc}
\title{robust retrieval of OHLC for a single pair}
\usage{
kraken_ohlc(pair, interval = 1440)
}
\arguments{
\item{pair}{assetpair to retrieve OHLC}

\item{interval}{integer corresponding to time interval}
}
\value{
OHLC dataframe
}
\description{
robust retrieval of OHLC for a single pair
}
