% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_ohlc_simple}
\alias{kraken_ohlc_simple}
\title{retrieval of OHLC data for a single pair}
\usage{
kraken_ohlc_simple(pair, interval = 1440)
}
\arguments{
\item{pair}{an assetpair on Kraken}

\item{interval}{an interval on Kraken}
}
\value{
OHLC dataframe
}
\description{
retrieval of OHLC data for a single pair
}
