% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_orderbook}
\alias{kraken_orderbook}
\title{Get order book}
\usage{
kraken_orderbook(pair, count = 10)
}
\arguments{
\item{pair}{asset pair to get market depth for}

\item{count}{maximum number of asks/bids (optional)}
}
\value{
dataframe of pair name and market depth
}
\description{
Get order book
}
