% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kraken_query}
\alias{kraken_query}
\title{Base functionality to call the Kraken API directly}
\usage{
kraken_query(url, args = NULL, private = FALSE, account = NULL)
}
\arguments{
\item{url}{Character input URL to query}

\item{args}{List of additional arguments to query}

\item{private}{Boolean value indicating a private or public query}

\item{account}{Pre-defined character input when using a private query}
}
\value{
Returns the result from the API query to Kraken
}
\description{
Base functionality to call the Kraken API directly
}
