% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_query_orders_info}
\alias{kraken_query_orders_info}
\title{Query orders info}
\usage{
kraken_query_orders_info(txid = NULL, userref = NULL)
}
\arguments{
\item{txid}{comma delimited list of transaction ids to query info about (20 maximum)}

\item{userref}{restrict results to given user reference id (optional)}
}
\value{
associative array of orders info
}
\description{
Query orders info
}
