% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_rebalance_portfolio}
\alias{kraken_rebalance_portfolio}
\title{Rebalance the portfolio on Kraken}
\usage{
kraken_rebalance_portfolio(current_allocation)
}
\arguments{
\item{current_allocation}{the current allocation}
}
\value{
nothing
}
\description{
Rebalance the portfolio on Kraken
}
