% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_servertime}
\alias{kraken_servertime}
\title{get Kraken servertime}
\usage{
kraken_servertime()
}
\value{
a unix timestamp representing Kraken's servertime
}
\description{
get Kraken servertime
}
