% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_spread}
\alias{kraken_spread}
\title{Get recent spread data}
\usage{
kraken_spread(pair)
}
\arguments{
\item{pair}{asset pair to get spread data for}
}
\value{
dataframe of pair name and recent spread data
}
\description{
Get recent spread data
}
