% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_ticker}
\alias{kraken_ticker}
\title{get a ticker}
\usage{
kraken_ticker(pair)
}
\arguments{
\item{pair}{an assetpair}
}
\value{
ticker information for a given assetpair
}
\description{
get a ticker
}
