% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_withdraw_funds}
\alias{kraken_withdraw_funds}
\title{Withdraw funds}
\usage{
kraken_withdraw_funds(aclass = NULL, asset, key, amount)
}
\arguments{
\item{aclass}{(optional) asset class}

\item{asset}{asset being withdrawn}

\item{key}{withdrawal key name, as set up on your account}

\item{amount}{amount to withdraw, including fee's}
}
\value{
associative array of withdrawal transaction:
}
\description{
Withdraw funds
}
