% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_candles}
\alias{nomics_candles}
\title{The exchange candles endpoint returns raw open, close, high, low, and volume information for Nomics Markets.
The data is not aggregated, therefore prices are in the quote currency of the market and volume is in the base currency of the market.}
\usage{
nomics_candles(interval, exchange, market, start = NULL, end = NULL)
}
\arguments{
\item{interval}{Time interval of the candle}

\item{exchange}{Exchange ID}

\item{market}{Market ID}

\item{start}{(optional)}

\item{end}{(optional)}
}
\value{
A list of candles
}
\description{
The exchange candles endpoint returns raw open, close, high, low, and volume information for Nomics Markets.
The data is not aggregated, therefore prices are in the quote currency of the market and volume is in the base currency of the market.
}
