% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_candles_aggr}
\alias{nomics_candles_aggr}
\title{The candles endpoint returns aggregated open, high, low, close, and volume information for Nomics currencies.
When asking for candles, a currency is provided as a parameter.
Nomics aggregates all markets where the given currency is the base currency
and the quote currency is a fiat currency, BTC, or ETH and returns all values in USD.
Candles are aggregated across all markets for the base currencies, which necessitates converting to a common quote currency.
Nomics converts all markets into USD in order to aggregated candles.}
\usage{
nomics_candles_aggr(interval = "1d", currency = "btc", start = NULL,
  end = NULL)
}
\arguments{
\item{interval}{(optional) Time interval of the candle ('1d' or '1h'). Default is '1d'.}

\item{currency}{(optional) Currency ID. Default is 'btc'}

\item{start}{(optional) Start time of the interval in POSIXct format. If not provided, starts from first candle.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
A list of candles
}
\description{
The candles endpoint returns aggregated open, high, low, close, and volume information for Nomics currencies.
When asking for candles, a currency is provided as a parameter.
Nomics aggregates all markets where the given currency is the base currency
and the quote currency is a fiat currency, BTC, or ETH and returns all values in USD.
Candles are aggregated across all markets for the base currencies, which necessitates converting to a common quote currency.
Nomics converts all markets into USD in order to aggregated candles.
}
