% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_currencies}
\alias{nomics_currencies}
\title{The currencies endpoint returns all the currencies that Nomics supports}
\usage{
nomics_currencies()
}
\value{
supported currencies in a tibble
}
\description{
The currencies endpoint returns all the currencies that Nomics supports
}
\examples{
\dontrun{
nomics_currencies()
}
}
