% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_currencies_interval}
\alias{nomics_currencies_interval}
\title{Open and close prices and volume for all currencies between a customizable time range.}
\usage{
nomics_currencies_interval(start, end = NULL)
}
\arguments{
\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Open, close, and volume for all currencies in a time range
}
\description{
Open and close prices and volume for all currencies between a customizable time range.
}
