% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_currencies_sparkline}
\alias{nomics_currencies_sparkline}
\title{The currencies sparkline endpoint returns prices for all currencies within a customizable time interval
 suitable for sparkline charts.
Note the timestamps and prices are built off of OHLCV candles using the close price.
This means that the timestamp represents the start of the candle, and the price is the close price of that candle.
This means the response's final timestamp and price value are always as current as possible,
 but also that the price is effectively "off" by one candle.
 This endpoint is designed to serve as a convenient way to render sparklines,
 if you need exactly aligned times and prices you can use the Aggregated OHLCV Candles endpoint.}
\usage{
nomics_currencies_sparkline(start, end = NULL)
}
\arguments{
\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Currency performance over time for all currencies over the requested time period.
}
\description{
The currencies sparkline endpoint returns prices for all currencies within a customizable time interval
 suitable for sparkline charts.
Note the timestamps and prices are built off of OHLCV candles using the close price.
This means that the timestamp represents the start of the candle, and the price is the close price of that candle.
This means the response's final timestamp and price value are always as current as possible,
 but also that the price is effectively "off" by one candle.
 This endpoint is designed to serve as a convenient way to render sparklines,
 if you need exactly aligned times and prices you can use the Aggregated OHLCV Candles endpoint.
}
