% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_exchange_market_interval}
\alias{nomics_exchange_market_interval}
\title{The exchange market interval endpoint returns a summary of information
about all markets over a configurable time interval in their native values.}
\usage{
nomics_exchange_market_interval(currency = NULL, exchange = NULL, start,
  end = NULL)
}
\arguments{
\item{currency}{(optional) Nomics Currency ID to filter by. If present, only markets with this currency as the base or quote will be returned.}

\item{exchange}{(optional) Nomics Exchange ID to filter by. If present, only markets on this exchange will be returned}

\item{start}{(required) Start time of the interval (as POSIXct).}

\item{end}{(optional) End time of the interval (as POSIXct). If not provided, the current time is used.}
}
\value{
A list of markets with price and volume information for a currency
}
\description{
The exchange market interval endpoint returns a summary of information
about all markets over a configurable time interval in their native values.
}
