% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_exchange_market_prices}
\alias{nomics_exchange_market_prices}
\title{The exchange market prices endpoint returns prices for the last trade in each market}
\usage{
nomics_exchange_market_prices(currency = NULL, exchange = NULL)
}
\arguments{
\item{currency}{Nomics Currency ID to filter by. If present, only markets with this currency as the base or quote will be returned.}

\item{exchange}{Nomics Exchange ID to filter by. If present, only markets on this exchange will be returned}
}
\value{
A list of markets with their prices
}
\description{
The exchange market prices endpoint returns prices for the last trade in each market
}
