% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_exchange_rates}
\alias{nomics_exchange_rates}
\title{The exchange rates endpoint returns the current exchange rates used by Nomics to convert prices from markets into USD.
This contains Fiat currencies as well as a BTC and ETH quote prices.
This endpoint helps normalize data across markets as well as to provide localization for users.}
\usage{
nomics_exchange_rates()
}
\value{
List of exchange rates into USD
}
\description{
The exchange rates endpoint returns the current exchange rates used by Nomics to convert prices from markets into USD.
This contains Fiat currencies as well as a BTC and ETH quote prices.
This endpoint helps normalize data across markets as well as to provide localization for users.
}
