% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_exchange_rates_history}
\alias{nomics_exchange_rates_history}
\title{Exchange rates for every point in a time range.
This endpoint can be used with other interval endpoints to convert values into a desired quote currency.
The currency parameter must be a Nomics Quote Currency, to get all Nomics Quote Currencies,
use the /exchange-rates endpoint for all current rates}
\usage{
nomics_exchange_rates_history(currency, start, end = NULL)
}
\arguments{
\item{currency}{(require) Currency ID}

\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Exchange rates for every point in a time range
}
\description{
Exchange rates for every point in a time range.
This endpoint can be used with other interval endpoints to convert values into a desired quote currency.
The currency parameter must be a Nomics Quote Currency, to get all Nomics Quote Currencies,
use the /exchange-rates endpoint for all current rates
}
