% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_interval_exchange_rates}
\alias{nomics_interval_exchange_rates}
\title{Exchange rates to convert from USD over a time interval.
This endpoint can be used with other interval endpoints to convert values into a desired quote currency.}
\usage{
nomics_interval_exchange_rates(start, end = NULL)
}
\arguments{
\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
List of exchange rates open and close information
}
\description{
Exchange rates to convert from USD over a time interval.
This endpoint can be used with other interval endpoints to convert values into a desired quote currency.
}
