% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_interval_markets}
\alias{nomics_interval_markets}
\title{The market interval endpoint returns a summary of information about all markets (in USD) based
in a given currency over a configurable time interval.}
\usage{
nomics_interval_markets(currency, hours = NULL, start = NULL,
  end = NULL)
}
\arguments{
\item{currency}{(required) Nomics currency id to query information for}

\item{hours}{(optional) Number of hours back to calculate date [1 ... 87600]. Default = 1}

\item{start}{(optional) Start time of the interval (as POSIXct). If not provided, it is computed using the hours parameter.}

\item{end}{(optional) End time of the interval (as POSIXct). If not provided, the current time is used.}
}
\value{
A list of markets with price and volume information for a currency.
}
\description{
The market interval endpoint returns a summary of information about all markets (in USD) based
in a given currency over a configurable time interval.
}
