% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_interval_mcap_aggr}
\alias{nomics_interval_mcap_aggr}
\title{Market capitalization history in aggregate}
\usage{
nomics_interval_mcap_aggr(start = NULL, end = NULL)
}
\arguments{
\item{start}{(optional) Start time of the interval in POSIXct format. If not provided, yesterday will be used.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Performance over time for the entire market.
}
\description{
\code{nomics_interval_mcap_aggr} returns the aggregate market cap for all cryptoassets
  at intervals between the requested time period.
}
