% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_interval_mcap_indiv}
\alias{nomics_interval_mcap_indiv}
\title{Market capitalization for individuals currencies historical}
\usage{
nomics_interval_mcap_indiv(start = NULL, end = NULL)
}
\arguments{
\item{start}{(optional) Start time of the interval in POSIXct format. If not provided, yesterday will be used.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\description{
\code{nomics_interval_mcap_indiv} returns the individual market caps for all cryptoassets
  between a start and end date.
}
