% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_interval_supplies}
\alias{nomics_interval_supplies}
\title{Open and close supply information for all currencies between a customizable time interval.}
\usage{
nomics_interval_supplies(start, end = NULL)
}
\arguments{
\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Open and close supply information
}
\description{
Open and close supply information for all currencies between a customizable time interval.
}
