% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_market_prices}
\alias{nomics_market_prices}
\title{The market prices endpoint returns prices in USD for the last trade in each market with the given base currency.}
\usage{
nomics_market_prices(currency)
}
\arguments{
\item{currency}{Nomics Currency ID of the desired base currency}
}
\value{
A list of markets with their price
}
\description{
The market prices endpoint returns prices in USD for the last trade in each market with the given base currency.
}
