% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_mcap_hist}
\alias{nomics_mcap_hist}
\title{MarketCap History is the total market cap for all cryptoassets at intervals between the requested time period.}
\usage{
nomics_mcap_hist(start, end = NULL)
}
\arguments{
\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Performance over time for the entire market.
}
\description{
MarketCap History is the total market cap for all cryptoassets at intervals between the requested time period.
}
