% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_prices}
\alias{nomics_prices}
\title{The prices endpoint returns current prices for all currencies. Prices are updated every 10 seconds.}
\usage{
nomics_prices()
}
\value{
A list of currencies with their price
}
\description{
The prices endpoint returns current prices for all currencies. Prices are updated every 10 seconds.
}
