% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_prices_exchange_markets}
\alias{nomics_prices_exchange_markets}
\title{Exchange market prices}
\usage{
nomics_prices_exchange_markets(currency = "btc", exchange = NULL)
}
\arguments{
\item{currency}{(optional) Nomics Currency ID to filter by. If present, only markets with this currency as the base or quote will be returned.}

\item{exchange}{(optional) Nomics Exchange ID to filter by. If present, only markets on this exchange will be returned}
}
\value{
A list of markets with their prices
}
\description{
\code{nomics_prices_exchange_markets} returns prices for the last trade in each market
}
\references{
http://docs.nomics.com/#operation/getExchangeMarketPrices
}
