% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_volume_global}
\alias{nomics_volume_global}
\title{Volume History is the total volume for all cryptoassets in USD at intervals between the requested time period.}
\usage{
nomics_volume_global(start = NULL, end = NULL)
}
\arguments{
\item{start}{(required) Start time of the interval in POSIXct format.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
Volume over time for all currencies
}
\description{
Volume History is the total volume for all cryptoassets in USD at intervals between the requested time period.
}
