% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_retracement}
\alias{plot_retracement}
\title{Plot Retracement Ratio}
\usage{
plot_retracement(data)
}
\arguments{
\item{data}{A dataframe with at least \code{date} and \code{price} columns.}
}
\description{
Visualize retracement ratio over time.
}
