% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{query}
\alias{query}
\title{API querying function}
\usage{
query(url, config, post_data)
}
\arguments{
\item{url}{a url}

\item{config}{a config file to use in the POST method}

\item{post_data}{additional post data to use in the POST method}
}
\value{
a response
}
\description{
API querying function
}
\keyword{internal}
