% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken}
\alias{query_kraken}
\title{Functionality to interface with Kraken API}
\usage{
query_kraken(url, args = NULL, private = FALSE, account = NULL)
}
\arguments{
\item{url}{Character input URL to query}

\item{args}{List of additional arguments to query}

\item{private}{Boolean value indicating a private or public query}

\item{account}{Pre-defined character input when using a private query}
}
\value{
Returns the result from the API query to Kraken
}
\description{
Functionality to interface with Kraken API
}
