% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_assetpairs}
\alias{query_kraken_assetpairs}
\title{Query Assetpairs available on Kraken}
\usage{
query_kraken_assetpairs(incl_darkpool = FALSE)
}
\value{
available assetpairs on Kraken
}
\description{
Query Assetpairs available on Kraken
}
