% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_ledger}
\alias{query_kraken_ledger}
\title{Query Kraken Ledger}
\usage{
query_kraken_ledger(data_wide)
}
\arguments{
\item{data_wide}{Historical dataframe in 'wide' format}
}
\value{
returns the ledger dataframe
}
\description{
Query Kraken Ledger
}
