% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_order}
\alias{query_kraken_order}
\title{Query a single order}
\usage{
query_kraken_order(order_id)
}
\arguments{
\item{order_id}{an orderid}
}
\value{
result for order
}
\description{
Query a single order
}
