% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_orderbook}
\alias{query_kraken_orderbook}
\title{Query orderbook for a given pair}
\usage{
query_kraken_orderbook(pair, count = 10)
}
\arguments{
\item{pair}{an assetpair}

\item{count}{number of lines to retrieve}
}
\value{
bid and ask orderbook
}
\description{
Query orderbook for a given pair
}
