% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_prices}
\alias{query_kraken_prices}
\title{Query OHLC data from Kraken}
\usage{
query_kraken_prices(pairs, interval = 1440)
}
\arguments{
\item{pairs}{list of assetpairs to retrieve OHLC data}

\item{interval}{integer corresponding to time interval}
}
\value{
a dataframe with OHLC data
}
\description{
Query OHLC data from Kraken
}
