% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_spread}
\alias{query_kraken_spread}
\title{Query spread information from Kraken}
\usage{
query_kraken_spread(pair)
}
\arguments{
\item{pair}{an assetpair}
}
\value{
dataframe with spread information
}
\description{
Query spread information from Kraken
}
