% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_ticker}
\alias{query_kraken_ticker}
\title{Query a ticker}
\usage{
query_kraken_ticker(pair)
}
\arguments{
\item{pair}{an assetpair}
}
\value{
ticker information for a given assetpair
}
\description{
Query a ticker
}
