% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{query_nomics}
\alias{query_nomics}
\title{Utility function to query Nomics API}
\usage{
query_nomics(endpoint, args = NULL)
}
\arguments{
\item{endpoint}{an endpoint}

\item{args}{a list of arguments to include in the API call}
}
\value{
parsed content from the API call
}
\description{
Utility function to query Nomics API
}
\keyword{internal}
