% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rebalance_kraken_portfolio}
\alias{rebalance_kraken_portfolio}
\title{Rebalance the portfolio on Kraken}
\usage{
rebalance_kraken_portfolio(current_allocation)
}
\arguments{
\item{current_allocation}{the current allocation}
}
\value{
nothing
}
\description{
Rebalance the portfolio on Kraken
}
