% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{secret_binance}
\alias{secret_binance}
\title{Look up Binance API secret stored in the environment}
\usage{
secret_binance()
}
\value{
string
}
\description{
Look up Binance API secret stored in the environment
}
\keyword{internal}
