% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{secret_kraken}
\alias{secret_kraken}
\title{Look up Kraken API secret stored in the environment}
\usage{
secret_kraken()
}
\value{
string
}
\description{
Look up Kraken API secret stored in the environment
}
\keyword{internal}
