% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{sign_kraken}
\alias{sign_kraken}
\title{Signature method for the Kraken API}
\usage{
sign_kraken(url, secret, key, nonce, post_data)
}
\arguments{
\item{url}{a url}

\item{secret}{API secret}

\item{key}{API key}

\item{nonce}{a nonce value}

\item{post_data}{additional post data to include in Signature}
}
\value{
header to include in the query_kraken call
}
\description{
Signature method for the Kraken API
}
\keyword{internal}
