% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_rfc3339}
\alias{to_rfc3339}
\title{Convert an R date to the RFC3339 format used in Nomics API}
\usage{
to_rfc3339(date)
}
\arguments{
\item{date}{a date in POSIXct format}
}
\value{
an rfc3339 representaton of date
}
\description{
Convert an R date to the RFC3339 format used in Nomics API
}
