% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptor.R
\docType{package}
\name{cryptor}
\alias{cryptor}
\alias{cryptor-package}
\title{cryptor: An R wrapper for the \url{https://www.cryptocompare.com} API.}
\description{
This package provides access to the crypto currency API documented at
\url{https://www.cryptocompare.com/api}. The API is currently public and can
be accessed without authentication or credentials. The API authors have enforced
call \href{https://min-api.cryptocompare.com/stats/rate/hour/limit}{hourly} and
\href{https://min-api.cryptocompare.com/stats/rate/second/limit}{second} limits.
}
\section{Functions}{

}

