% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_api_limit}
\alias{get_api_limit}
\title{Get number of API calls used and remaining in last hour or last second.}
\usage{
get_api_limit(time = "hour")
}
\arguments{
\item{time}{character. Either "hour", "minute", or "second". Defaults to "hour".}
}
\value{
A single tibble with two rows - one for calls made and one for calls
remaining
}
\description{
\code{get_api_limit} returns a tibble detailing API calls made and API calls
remaining against established rate limits.
}
\examples{
\dontrun{
# Get hour limit
get_api_limit()

# Get second limit
get_api_limit(time = "second")
}
}
\references{
\url{https://www.cryptocompare.com/api#requests}
}
