% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_exchanges}
\alias{get_exchanges}
\title{List all available exchanges.}
\usage{
get_exchanges(sign = FALSE, app_name = NULL)
}
\arguments{
\item{sign}{logical. Should the server sign the response? Defaults to FALSE.}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing details for each exchange available through the API.
}
\description{
\code{get_exchanges} provides details on each exchange available through the
API
}
\examples{
\dontrun{
exchanges <- get_exchanges()
head(exchanges)
}

}
\references{
\url{https://min-api.cryptocompare.com}
}
