% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_api.R
\name{get_top_exchanges}
\alias{get_top_exchanges}
\title{Get top exchanges by volume for a given currency pair.}
\usage{
get_top_exchanges(fsym, tsym, limit = 5, sign = FALSE, app_name = NULL)
}
\arguments{
\item{fsym}{character. 3 letter name for coin to retreive price history for.}

\item{tsym}{character. 3 letter name for how price should be reported for
\code{fsym}.}

\item{limit}{numeric. Maximum number of returned exchanges. Between 1 and 50.}

\item{sign}{logical. Should the server sign the response? Defaults to FALSE}

\item{app_name}{character. Name of app to be passed in API request. Defaults
to NULL.}
}
\value{
A tibble containing volume details for the top \code{limit} exchanges.
}
\description{
\code{get_top_exchanges} returns a tibble containing details about the top
exchanges by volume for a given currency pair.
}
\examples{
\dontrun{

}

}
\references{
\url{https://min-api.cryptocompare.com}
}
