% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{kraken_assetpairs}
\alias{kraken_assetpairs}
\title{get available assetpairs}
\usage{
kraken_assetpairs(excl_darkpool = TRUE)
}
\value{
available assetpairs on Kraken
}
\description{
get available assetpairs
}
