% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_assetpairs}
\alias{kraken_assetpairs}
\title{Get available assetpairs}
\usage{
kraken_assetpairs(excl_darkpool = TRUE)
}
\arguments{
\item{excl_darkpool}{(optional) exclude darkpool assetpairs from the list (default = TRUE).}
}
\value{
available assetpairs on Kraken
}
\description{
Get available assetpairs
}
