% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{kraken_rebalance_portfolio}
\alias{kraken_rebalance_portfolio}
\title{Rebalance the portfolio on Kraken}
\usage{
kraken_rebalance_portfolio(allocation_summary)
}
\arguments{
\item{allocation_summary}{the current allocation (returned from kraken_allocation_current())}
}
\value{
void
}
\description{
Rebalance the portfolio on Kraken
}
