% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_candles}
\alias{nomics_candles}
\title{Exchange candles endpoint}
\usage{
nomics_candles(interval = "1d", exchange, market, start = NULL,
  end = NULL)
}
\arguments{
\item{interval}{(optional) Time interval of the candle ('1d' or '1h'). Default is '1d'.}

\item{exchange}{Exchange ID}

\item{market}{Market ID}

\item{start}{(optional) Start time of the interval in POSIXct format. If not provided, starts from first candle.}

\item{end}{(optional) End time of the interval in POSIXct format. If not provided, the current time is used.}
}
\value{
A list of candles
}
\description{
returns raw open, close, high, low, and volume information for Nomics Markets.
  The data is not aggregated, therefore prices are in the quote currency of the market and
  volume is in the base currency of the market.
}
