% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_currencies}
\alias{nomics_currencies}
\title{Nomics currencies}
\usage{
nomics_currencies()
}
\value{
a tibble
}
\description{
\code{nomics_currencies} returns all the currencies that Nomics supports.
}
\examples{
nomics_currencies()

}
\references{
http://docs.nomics.com/#operation/getCurrencies
}
