% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nomics.R
\name{nomics_markets}
\alias{nomics_markets}
\title{The markets endpoint returns information on the exchanges and markets that Nomics supports,
in addition to the Nomics currency identifiers for the base and quote currency.}
\usage{
nomics_markets(exchange = NULL, base = NULL, quote = NULL)
}
\arguments{
\item{exchange}{(optional) a string of exchange ID to filter by (e.g. "kraken")}

\item{base}{(optional) a comma separated list of base currencies to filter by (e.g. "BTC,BCH" / "ETH")}

\item{quote}{(optional) a comma separated list of quote currencies to filter by (e.e. "ETH,ETC" / "BCH")}
}
\value{
a dataframe of markets
}
\description{
The markets endpoint returns information on the exchanges and markets that Nomics supports,
in addition to the Nomics currency identifiers for the base and quote currency.
}
