% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_kraken.R
\name{query_kraken}
\alias{query_kraken}
\title{Base functionality to call the Kraken API directly}
\usage{
query_kraken(url, sign = FALSE, args = NULL)
}
\arguments{
\item{url}{Character input URL to query}

\item{sign}{Boolean indicating if the request is private or public}

\item{args}{List of additional arguments to query}
}
\value{
Returns the result from the API query to Kraken
}
\description{
Base functionality to call the Kraken API directly
}
