% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_ohlc}
\alias{query_kraken_ohlc}
\title{(internal) get ohlc data}
\usage{
query_kraken_ohlc(pair, interval)
}
\arguments{
\item{pair}{assetpair to retrieve OHLC}

\item{interval}{integer corresponding to time interval}
}
\description{
(internal) get ohlc data
}
