% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_kraken_order}
\alias{query_kraken_order}
\title{get a single open order by order_id}
\usage{
query_kraken_order(order_id)
}
\arguments{
\item{order_id}{an orderid}
}
\value{
result for order
}
\description{
get a single open order by order_id
}
